Changed Recipeoverview to RecipeDetail Card displays all the details
Not the most beatifull UI.  Also We have to delete all the recipes that are not complete. You have to click on a liked recipe to see the details. When you click on a recipe that does not has any ingredients or not a array of ingredients it will give an error recipe.ingredients.map is not a function.